requirements.txt
torch==1.13.0
librosa
cython==0.29.35
transformers==4.29.2
datasets==2.10.1
webdataset
black
scipy
google-cloud-storage==2.10.0
git+https://github.com/CPJKU/madmom.git
dill==0.3.1.1
numpy
pandas
ffmpeg-python
beautifulsoup4
msgspec
sentencepiece
